WIP: Started debugging situations where the serialization crashes.
This seems to be at least partially due to a bug in MPS which causes large
allocations to succeed even though they should not. This bug seems to be due to
an integer overflow error in pointer arithmetics in mps_ap_fill, but also when
computing addresses when MPS is handling the fill. This error is likely not the
cause of crashes on 64-bit systems, as pointers are wider there, but sizes from
the serialization are still 32-bit.

Another issue that is present is that we may overflow in our size calculations
in some cases, and return an allocation size that is too small. We should check
this when allocating arrays of things (this is likely not an issue with
fixed-size allocations, as we would run out of available memory way before we
generate such object descriptions).