Fixed crashes in serialization found so far.
The crash was that the MPS writes to constant memory in some cases. In
particular, when the MPS has allocated large amounts of memory, it passes null
pointers to FIX2, which always writes to pointers. This is fine in and of
itself, but as some GcType instances are allocated in read-only memory, it will
make the MPS crash. All of these have their type set to null (out of necessity,
as we don't have addresses to type objects at compile-time), so we can identify
them that way.