Fixed a bug in WeakSet that could cause a crash.
If remove was called when the WeakSet had become empty due to objects no longer
being alive, the WeakSet's data would be cleared but then accessed. Now this
happens in the right order, so that the WeakSet no longer tries to remove
elements from an empty set in this situation.