Errors found by the model checker are shown properly.
Previously, in cases where a data race occurred in conjunction with a barrier,
the model checker would emit a sequence where the error was missed in the
visualization. This was due to an error in the handling of barriers. If a
statement contained a piece that would cause a data race before the barrier, the
barrier step would stop right before the operation, while a "regular" step would
stop right after the operation was completed, thus clearing the state. Now, we
make sure to check for data races right before barriers regardless of how the
code arrived there.