Fixed an issue in Progvis:
When a thread was woken from sleep, or for some other reason had to revert from
a barrier step to a "normal" step, it was possible for the logic that stopped a
thread after a release barrier to not trigger.

As this state was essentially hidden from the model checker, it could lead to
errors indicating that the model checker failed to advance the program to the
desired state.