Re-worked the icache management for ARM64.
The new scheme is based on the examples in the ARM Architecture Reference, where
it is possible to invalidate the icace for all cores from the core that actually
performs the modifications. All of this can be done in userspace.

The MPS has also been modified to allow calling the cache invalidation logic in
all threads when they have been paused. As the example from the ARM manual does
not require other threads to know exactly what has been modified, this works
well without any extra data structures.

The only thing that might remain is invalidating the icache on other CPUs
whenever we create new code. I think this is not necessary: if code is allocated
in previously unused memory, that memory is not cached by other cores. If it was
used, but has recently been freed, they would have executed a clear cache
instruction whenever the MPS moved the memory.