Fixed handling of results on Arm64.
Also make sure to sort SimpleDesc according to offset. Otherwise the Arm64
backend will fail to detect arrays of types.

With this change, Progvis seems to work on Arm64.

What remains is to refactor the X64 backend to use the new Params class, and to
figure out why Storm crashes on Arm64 in some cases. This is most likely due to
a memory corruption issue somewhere.