Likely fixed the issue with random crashes on Arm64.
The clearICache call in the format scanner accidentally scanned one word too
far. This was fine most of the time, but could lead to an access in a protected
page. This was rare since this situation required that a code allocation ended
up tightly to a page boundary, and that MPS decided to protect the next page.