Don't use markers in the Emacs plugin. That makes it misbehave.
Otherwise, if one edit overlaps with another, the offsets will be incorrect,
which in turn causes the color information to not match later on.

For example: in a buffer, insert a newline, go up one line and insert another
newline. If old color information is replayed at this point, the information
will be offset by one character in after the inserts.

As an additional bonus, this means that we don't have to store as many markers
in Emacs. This is good as the documentation states that too many markers will
reduce performance.