Ensured atomicity of iteration in the InlineSet.
This was important since an assertion iterated through the set outside the
context of the garbage collector. As such, it failed randomly, even though it
should maybe not do that.

Also, demoted the assertion to a dbg_assert, since iterating the InlineSet is
quite expensive if many UThreads are active.