Updated the UThreads library to work when the SEHOP mitigation is active.
It is not enabled by default on desktop versions of Windows, but might be
activated on managed installs, and on Windows Server.

The code here always adds a particular SEH frame on Windows. On systems where
SEHOP is present, this works regardless of whether the SEHOP is used or not. It
might fail on systems earlier than that, if the last SEH handler is ever
called (which is something that should never happen -- if it happens, the
process would likely crash anyway). According to MSDN documentation, it seems
like SEHOP was introduced in Windows Vista. Since Storm only really supports
systems from Windows 7 and upwards (the Gui) library, this should not be an
issue.