Fixed layout issue in Progvis.
If a (docked) thread would be created in the same step as an allocation, the
algorithm would fall back to the default position of (0, 0), which did not
respect the 'layoutEdgeSpace' variable as one would expect. Now, the algorithm
both handles the creation of new threads, and updates the default to respect the
edge spacing.