Replaced usage of IStream::read() with IStream::readAll() where applicable.
This fixes bugs where various parts of the system erroneously thought that the
end of stream was reached whenever IStream::read() would not fill the buffer
entirely. This is not correct, as IStream::read() is free to read as little data
as it wishes. The function IStream::readAll() should be used to coalesce reads.