The 'as' operator now works for Variant types as well.
This required re-structuring how Conditions behave. Now, they get a chance to
execute code in the true path of the conditonal to initialize the variable. This
means that the scoping of the created variable is more accurate (it does not get
destroyed if it was not initialized), and the Maybe casts generate less code.