Renamed 'Value::canStore' to 'Value::mayReferTo'.
This is to differentiate between ability to store an instance of a type, and the
ability to refer to types (by reference, through conversions). 'canStore' is
more about references, and thus operates on a more abstract level. The new
function 'mayStore' operates storage-wise, and is thus more strict regarding
values.

This change is in preparation for clarifying the semantics for creating
variables of base types, but that are initialized with derived types in Basic
Storm.