Fixed off-by-one in exception handling in 64-bit backends.
Previously, variable activations would not be respected until one instruction
later than their appearance. This was due to how function calls work, the
instruction referred in the backtrace is the next instruction to execute, and
not the faulting instruction. As such, we pad call instructions with nop
instructions in cases where this distinction matters.