Changed license on Storm to the 2-clause BSD license.
Storm and all associated core libraries are now 2-clause BSD. This change was
made partially in response to discussions about linking OpenSSL to software
covered by GPL licenses, but also to make it clear that it is possible to link
parts of the Storm runtime into a coherent binary (whenever that will be
possible) without worrying about needing to be able to replace the parts covered
by LGPL without all the source code.

All authors of the SQL library have been consulted about the change in license,
and they have all responded positively to the change through various channels.