Avoid copying too much to memory after function calls on X64.
Previously, when returning types with a size of 2 bytes, the backend would
generate code that wrote an entire 32-bit address to memory. This would result
in the possibility of overwriting other pointers on the stack. Now, the
situation is detected and handled properly (bytewise writes + shifts).

This is not an issue on ARM or 32-bit systems, since the ARM backend forces
pointer alignment of all variables, and the X86 backend would pass such values
in memory anyway.