Fixed a bug where accessing variables in the range-for loop could cause a stack overflow.
As a side-effect, this introduces a "variableHere" function to the Block in
Basic Storm that only examines variables in the current block. This is used by
the name-tree based lookup now, which reduces the number of duplicate hash
lookups performed when looking up variables. As such, this should make Basic
Storm slightly faster, especially for nested blocks.