Reworked how exports are handled to avoid surprises.
Exports are now handled when adding things to ScopeLookup objects, rather than
modifying the "find" member of "Package". This means that exports will not be
visible in the package that exports the thing. For example:

Package "a" exports "b".

Previously, all members of "b" would be visible to code inside of "a" as well,
which was non-intuitive. Now, this is not the case since "a" does not import
"b". However, all files that import "a" will also automatically import "b".