Relaxed limitations on actors.
It is now possible to define an actor on any thread, and then let derived
classes restrict it to a particular thread later on in the hierarchy. For
example:

class A on ? {}

class B : extends A, on Compiler {}

This can be used to make the syntax language more thread-agnostic in the future.