Fixed a dependency issue introduced in 12fc3b636e5ebfb3b6d6ebaf26fd4440dd3cb1f1
It caused 'loadAll' to be called too early in some cases, meaning that the code
in 'loadAll' made decisions based on incomplete inheritance- and threading
information.

Fixed by both making 'loadAll' call 'lookupTypes', and by making all type
lookups happen in a scope relative to the parent of the type, rather than the
type itself. Having them relative to the type itself makes the second
lookup (i.e. looking for a thread) look inside the super-class if specified,
which causes it to load too early.

This change likely makes the 'allowLazyLoad' variable unnecessary. It is
retained in case the user does something unexpected (similarly to the code that
ensures 'lookupTypes' is called from 'loadAll').